on avance un peu

+++ libft/ft_atoi.c
@@ -6,18 +6,18 @@
 /*   By: ajearuth <ajearuth@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/21 11:38:45 by ajearuth          #+#    #+#             */
-/*   Updated: 2021/05/23 15:35:10 by ajearuth         ###   ########.fr       */
+/*   Updated: 2021/09/22 17:33:48 by ajearuth         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+uint64_t	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	nb;
+	uint64_t	i;
+	uint64_t	sign;
+	uint64_t	nb;
 
 	i = 0;
 	sign = 1;
